fixed floating point exception

+++ pop_nucdiv.cpp
@@ -255,7 +255,7 @@ void nucdivData::calc_nucdiv(void)
 		{
 			sum = 0.0;
 			for (k=0; k < segsites; k++)
-				sum += (freq[i][k] * (ncov[j][k] - freq[j][k]) + freq[j][k] * (ncov[i][k] - freq[i][k])) / (ncov[i][k] * ncov[j][k]);
+				sum += (double)(freq[i][k] * (ncov[j][k] - freq[j][k]) + freq[j][k] * (ncov[i][k] - freq[i][k])) / (ncov[i][k] * ncov[j][k]);
 			pib[UTIDX(sm->npops,i,j)] = sum / ns_between[UTIDX(sm->npops,i,j)];
 		}
 	}
